Desktop/StreamCAG/requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
datasets>=2.14.0
accelerate>=0.21.0
bitsandbytes>=0.41.0
gradio>=3.44.0
python-dotenv>=1.0.0
